Replace MonadZAuth with Input ZAuthEnv and Embed IO
Another way to do this would've been to implement stuff in the prod interpreter
of AuthenticationSubsytem using MonadZAuth and then get rid of MonadZAuth. But
it so happened this way. This way is worse because it has caused a lot of
changes to constraints in brig which would need to be undone. Perhaps that
happens in the next commit.