Change saml2-web-sso to only use verified signed assertions.
This is a bit large because the old code also did (mostly redundant,
and always misleading) saml data validation on unsigned parts of the
authentication response (as required in the standard).  The new code
doesn't do this any more, which caused a lot of changes where
AuthnReponse is changed by a non-empty list of assertions.

saml2-web-sso compiles, but a lot of unit tests are failing.