AuthenticationSubsystem: Move newCookie into it (#4479)
* integration: Move testInvalidCookie from brig-integration

* Move Brig.ZAuth to Wire.AuthenticationSubsystem.ZAuth

This needs more changes to the zauth library before it can be integrated in the
the AuthenticationSubsystem.

* zauth: Polysemize token creation and use it in AuthSubsys.ZAuth

* Remove TokenPair type class

The 'newAccessToken' function is moved to the 'AccessTokenLike' typeclass
returns an 'AccessToken' type which contains the serialized token, removing the
need for polymorphism.

Similarly 'renewAccessToken' function in the 'UserTokenLike' typeclass returns
an 'AccessToken'.

We might be able to get rid of all of this polymophism as things are not _so_
complicated. This is just the first step.

* Remove accessTokenClient from AccessTokenLike

It was only used in integration tests and only for one type of access token

* zauth: Unify User/LegalHoldUser and Acess/LegalHoldAccess using a phantom type paramter

These are essentially same things and we have invented type classes to treat
them so, but if they contain the same data structures inside and have same
serialisation/de-serialisation logic, there is little point in having new type
wrappers for LegalHoldUser and LegalHoldAccess.

The phantom type parameter is of type `UserTokenType` with values `ActualUser`
and `LHUser`. These names don't completely make any sense because there is no
such thing as a `LegalHoldUser` different from `ActualUser`. It _should_ rather
be called something like `TokenType`, but that name is taken by something else.
I suspect this type will go away soon, so while transitioning its probably fine
to keep it as it.

Doing all this makes it easier for The typeclasses `AccessTokenLike` and
`UserTokenLike` to be merged. The actual use of those type classes is to
determine which ttls to apply to these the `UserTokenType`, which can now be
made clearer in the code.

* zauth: Delete unused validation functions

* Introduce mappings from UserTokenType to Token.Type

These help de-dupe bit more code.

* zauth: Polysemize validation

Delete bunch of unused code.

* zauth: Remove lenses and renamed constructors

* zauth: Use 'Type' as the parameter for the 'Token' type instead of body

Create a type family 'Body t' which defines the body based on the type of the
token. This removes illegal states like token of type Bot having a User as the
body.

As a result the token creation can be greatly simplified with it being passed
the parameters for creating the header and the body of correct type. This helps
reduce some duplication between ZAuth.Creation and
Wire.AuthenticationSubsystems.ZAuth. This also pushed the need for the random
number generation to the subsystems code making the creation even lighter. Maybe
we can get rid of polysemy from there too making the whole thing very simple.

The classes 'UserTokenLike' and 'AccessTokenLike' also become much simpler as
they can just have a general constraint like '(Body t ~ User)' and '(Body t ~
Access) respectively. Perhaps these classes can be gotten rid of once we can use
some state from the AuthenticationSubsystem.

* integration tests: Adjust expectations of reading invalid tokens

Since the token is generic now, read errors don't specify which type of token is
invalid to read. This only changes the message of the error and not the label,
so it is an acceptable change.

* AuthSubs.ZAuth: deduplicate mkUserToken' and mkLegalHoldUserToken

They're same functions now, this deals with a futurework

* Unify newAccessToken and newLegalHoldAccessToken

It is a bit reminiscent of the recently removed TokenPair class. Perhaps this
was we can merge AccessTokenLike and UserTokenLike soon.

* Unify newUserToken and newLegalHoldUserToken

* zauth: Implement renewToken in terms of newToken

The implementation only depended on newTokenImpl anyway.

* zauth: Remove polysemy

Every place it was used, it was immediately run. So there is little point in
keeping this abstraction around.

* Take newSessionToken out of UserTokenLike

This makes UserTokenLike indepedent of MonadZAuth, freeing it to be useable in a
polysemy action.

* Replace MonadZAuth with Input ZAuthEnv and Embed IO

Another way to do this would've been to implement stuff in the prod interpreter
of AuthenticationSubsytem using MonadZAuth and then get rid of MonadZAuth. But
it so happened this way. This way is worse because it has caused a lot of
changes to constraints in brig which would need to be undone. Perhaps that
happens in the next commit.

* Make wire-api dependent on polysemy-wire-zoo instead of the other around

This will help polysemy-wire-zoo to be usable in zauth

* zauth: Reintroduce polysemy at a lower level

This helps get rid of IO from token creation and validation, which would help
avoid IO from AuthenticationSubsystem.Interpreter, which is required because
MiniBackend doesn't run IO right now.

brig doesn't compile right now, but the functions will be moved into
AuthenticationSubsystem anyway, so playing type-tetris isn't as productive.

* AuthenticationSubsytem: Move code for issuing a new cookie

* Propagate constraint changes and newCookie moving to brig

* Don't use singletons.

The old code is good enough, and singletons may add more complexity
than they abstract away.

* Changelog.

* integration/testCookieLimit: Add threadDelay to get around failure

Brig updates its time every second. If we create many cookies very quickly all
of them get assigned the same cookie creation time, this makes it impossible to
guess which cookie will get evicted when we have too many cookies. So waiting a
second between cookie creation gets around this problem.

* AuthenticationSubsystem: Consolidate Input effects

* brig: Remove `Member (Input Env)` from everywhere

Move the required settings to AuthenticationSubsystemConfig.

* brig: Remove ZAuthEnv as an explicit Input

* Add test for randomConnId

---------

Co-authored-by: Matthias Fischmann <mf@zerobuzz.net>
Co-authored-by: Leif Battermann <leif.battermann@wire.com>