Migrations: Make (team_id, id) the primary key of user_group table
Since a group exists within bounds of a team, it will always be retrieved like
that. So it makes sense to first index on the team id. Sticking to semantics for
the sake of it might not always be the best thing to do, but here it will
actually make the queries faster.