cannon: Replace message count mechanism with detecting the end of initial sync (#4631)
* cannon: Replace message count mechanism with detecting the end of initial sync

The message count recieved in the queue info gotten as response of declaring the
queue may include some expired messages which have not yet made it to the head
of the queue. This makes the message count mechanism useless to the clients for
detecting the end of initial sync.

This commit detects end of initial sync based on sending a transient
notification in the queue. If we recieve this notification back, it would mean
that the queue is empty and the initial sync is complete.

The timing of sending this notification must be carefully chosen, it is sent in
one of these two conditions:

  1. If the message count in the queue info is 0, the queue must be empty as the
  expired messages are already considered expired when they reach the head of
  the queue. So its a good time to publish our end-of-initial-sync notification.
  If the queue is indeed empty we'll recieve it and forward it to the client,
  letting it know that it can consider itself up to date.

  2. If number of unacked messages becomes 0 right after recieving an
  acknowledgement and the initial sync hasn't already completed, this could mean
  either the queue is empty or there is some delay in receiving the next message
  from RabbitMQ. Here we publish our end-of-initial-sync notification. If there
  was a delay in receiving the next message from RabbitMQ, the consumer thread
  will increase the count for unacked messages and we'll get another opportunity
  to check this. Otherwise, the consumer thread will recieve this
  end-of-initial-sync notification and forward it to the client.

* changelog

* cannon: Slightly optimize end-of-initial-sync detection

* cannon: Use atomicModifyIORef' instead of modifyIORef'

The IORef is being manipulated from two different threads, so it should use the
atomic version of the function.

* cannon: Undo small part of optimisation

Description in the comment

* use latest unacked delivery tag instead of a counter

* insert sync message on websocket connection

* charts/cannon: Allow configuring rabbitMsMaxConnections and rabbitMqMaxChannels

These option were implemented previously but we forgot to add them in the helm
chart.

* cannon: Queue synchronization message as Persistent

Use configurable TTL with default value set to 28 days like for normal
notifications.

* cannon: Only add sync_marker to API version v9

* integration: Rename testQosLimit to testPrefetchCount

* changelog: config opt for cannon

* integration: createEventsWebSocketWitSync creates a random marker

Instead of expecting it to be passed in

---------

Co-authored-by: Leif Battermann <leif.battermann@wire.com>