integration-setup: Add explicit helm login to docker hub  (#4668)
* integration-setup: Add explicit helm login to docker hub

This is required because we have to run `helm dep update` manually before
running `helmfile` even if we run `helmfile` without `--skip-sync` because `helm
dep update` needs to happen in a certain order to make sure all nested local
dependencies are packaged correctly.

* helmfile: Explicitly login to docker hub

This seems necessary when downloading helm charts.

* Makefile: Mark helm-oci-login as phony