Heal detached background-worker consumers symptoms in tests  (#4634)
Actually, there isn't much we can do when a queue is externally deleted:

- There is a loop which handles the reconfiguration according to existing (!)
- queues Older federation backend versions (V0 and V1) cannot be changed
- anymore.

To heal the symptoms of detached consumers, I've decreased the cleanup interval
for local integration test runs to one Second. That should be a good tradeoff
  between the time to re-execute tests and not making to many bogus requests to
  RabbitMQ.

I've also added a test to ensure that reestablishing the consumer works.