Fix: IdP deletion after issuer update
When an Issuer gets replaced by an IdP update, it is kept in the
oldIssuers list for this IdP. But, it doesn't reference an active IdP
anymore.

Thus, looking up IdPs by this Issuer is deemed to fail.

Previously, deleting IdPs with such updated Issuers failed
(erroneously). This commit fixes this behaviour by acknowledging the
fact, that an old issuer may not point to an active IdP.