ConversationStore.Cassandra: Always cleanup deleted convs
Previously this was done in the store interpreter when getting one conv, but
done in galley code when getting a list of convs.

The cleanup logic is required only in Cassandra because it cannot guarantee that
deletions of rows from the member table along with rows from the conv table. So
the usual deletion logic is to mark the conv deleted first and then try to
delete the rows. This will all not be required in Postgres because we can do
these things with transactions.