ConversationStore.Postgres: Give up on optimizing the pagination query
The pagination can be done with a UNION query (and not a join as the comment
suggests), but it causes weird problems due to domain being NULL. We could
send the local domain in to always get the domain, but this just causes more
work for Postgres. Its probably more efficient to keep what we have.