Implement Postgres interpreter for ConversationStore (#4764)
* Extract actions for MLS commit lock in a separate effect

This will allow us to migrate rest of the conversation data to postgres

* ConversationStore: Remove action createConversationId

It has nothing to do with the store, the one place which calls this can call
`Random.uuid` instead.

* Extract logic to convert NewConversation -> StoredConversation outside the Casssadnra interpreter

* ConversationStore: Remove polymorphism from CreateMembers

It was used only twice seems like unnecesary leak of business logic into the
store effect.

* PostgresUnmarshall: Add a few instances

* ConversationStore: Delete unused Get(Remote,Local)MembersByDomain

* ConversationStore.Cassandra: Make functions for making conv objects from db more general

* ConversationStore.Cassandra: Always cleanup deleted convs

Previously this was done in the store interpreter when getting one conv, but
done in galley code when getting a list of convs.

The cleanup logic is required only in Cassandra because it cannot guarantee that
deletions of rows from the member table along with rows from the conv table. So
the usual deletion logic is to mark the conv deleted first and then try to
delete the rows. This will all not be required in Postgres because we can do
these things with transactions.

* galley: Allow selecting postgrsql for storing conversations

* ConversationStore: Rename CreateSubConversation to UpsertSubConversation

Also change the Postgres interpreter to do so. In the galley code, this is
explicitly being used as an Upsert.

* CovnersationStore: Add action GetConversationIds

This replaces the use of ListItems.

* CovnersationStore: Add action GetLocalConversationIds

This replaces the use of ListItems with LegacyPaging

* ConversationStore: CreateMembersInRemoteConversation -> Upsert...

* ConversationStore: CreateMembers -> UpsertMembers

* ConversationStore: CreateConversation -> UpsertConversation

* galley-integration: Ignore order of self conversation ids returned from /conversations/list-ids

---------

Co-authored-by: Matthias Fischmann <mf@zerobuzz.net>