ConversationStore: Implement migration to postgres (#4810)
* ConversationStore.Migration: Implement function to migrate a conversation

* ConversationStore.Cassandra: Add hybrid interpretter, which will work during migration

Exceptions:
1. Getting a paginated list of qualified conv Ids
2. It will not work well if the migration fails to delete conv data from
Cassandra after copying it to postgres

These problems will be solved in following commits

* ConverastionStore.Cassandra: Fix edge case bug in getLocalConvIds

When fetching `maxIds + 1` convs, it can happen that a user has exactly those
many convs left, in this case the `hasMore` field of the page would be false,
but we'd be sending a truncated list of convs.

* ConverastionStore: Make GetConverastionIds work during migration

This is done by making the pagingState encode last conversation Id served. The
store effect has a new action to list only remote conv ids. The
`GetConverastionIds` action has been removed and implemented generally using
`GetLocalConverastionIds` and `GetRemoteConversationIds`.

This makes `MultiTabelPage` type obsolete for conv ids, but its still kept
around so we don't break any APIs.

* ConversationStore.Cassandra: Save users joining their first remote conv in postgres

This is consistent with creating new conversations in postgres. This way when
the migration is complete already running galley instances won't create more
data in Cassandra

* ConversationStore.Cassandra: Use `embedClient`

* ConversationStore.Migration: Implement function to migrate a remote statuses of a user

* ConversationStore.Migration: Add top level functions to do the migration

* Conversation.{Migration,Cassandra}: Ensure pending deletes from Cassandra are never read

Exception: Listing local conversation ids and listing team conversation ids

* ConverastionStore.Cassandra: Document limitation in listing conv ids

* galley: Allow choosing migration interpreter for ConversationStore

* background-worker: Integrate with ConvSubsystem to allow migrating convs to postgres

* integration: Set cassandra keyspace and pg db names for dyn background worker

* integration: Add a test to test conversation migration

* background-worker: Add some observability to pg migration

* integration: Add another test with more convos, but use proteus for speed

* ConversationStore.{Cassandra,Postgres}: Use same ordering of UUIDs

Cassandra cares for UUID version in ordering. Postgres doesn't, so the Postgres
query needs to be weird.

* wire-subsystems: Move PGConstraints to Wire.Postgres

* ConversationStore.Cassandra: Implement SearchConversation for the migration interpreter

It returns empty.

* galley-integration: Change assertions about pagination

Because sometimes we don't have to get an empty page to realize we're at the end

* integration-setup: More CPU/memory for postgresql

Also stop any throttling, OOMKilling

* charts/integration: Mount background-worker secrets

* integration: Set cassandra keyspace for dynamic cannons

* integration-setup: Allow tests to run for 5 more mins

* galley-integration: Produce better error message

* docs: Add steps for migrations

* galley-integration: Relax the requirement that getConvs returns convs in same order as IDs